requirements.txt
diffusers
Pillow
torch
tqdm
